added motor superclass
