+++ src/beachbot/manipulators/vrepmotorsim.py
@@ -1,11 +1,12 @@
 from ..utils.vrepsimulation import vrep
+from beachbot.manipulators.motor import Motor
 
-class VrepMotorSim():
+class VrepMotorSim(Motor):
     def __init__(self, vrep_sim, motor_name, max_velocity=5):
+        super().__init__(motor_name)
         self.vrep_sim = vrep_sim
-        self._motor_name = motor_name
         self._max_velocity=max_velocity
-        self._motor_id = self.vrep_sim.getObject("/"+motor_name)
+        self._motor_id = self.vrep_sim.getObject("/"+self.name)
 
     @vrep    
     def change_speed(self, speed: int) -> None:
@@ -18,10 +19,8 @@ class VrepMotorSim():
                          positive values represent forward motion,
                          and zero stops the motor.
         """
-        if not -100 <= speed <= 100:
-            raise ValueError(
-                f"Speed must be between -100 and 100 inclusive. Received: {speed}"
-            )
+        super().change_speed(speed)
+
         # start moving at 15% power (as real platform has friction), if lower than 15% -> do not move
         if speed<15 and speed>-15:
             speed=0
@@ -32,6 +31,4 @@ class VrepMotorSim():
 
         self.vrep_sim.setJointTargetVelocity(self._motor_id, self._max_velocity * speed / 100)
 
-    def cleanup(self):
-         pass
